+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-### CHANGE LOG
+## CHANGE LOG
 
-#### v0.1.1
+### v0.1.1
 > Adding a new function `trackPageview` for GA pageview tracking.
+
+### v0.1.0
+> Updating the `react-ga` package.
+> Updating webpack process for building.